meta tags fix and screenshot
