+++ components/constants/header.tsx
@@ -1,0 +1,5 @@
+export const HEADER_TITLE: string = 'Woolball a name system with links'
+export const HEADER_DESCRIPTION: string =
+  'Woolball is a name system that let you save connections between names on the blockchain in the form of links. With Woolball links you can easily build social Daaps, add Web3 accounts to an existing Daaps, or implement a reputation system for names.'
+
+export const HEADER_IMAGE: string = 'https://woolball.xyz/images/screenshot.jpg'
